add NLP new thread

+++ src/openai_parser.py
@@ -6,7 +6,17 @@ from openai import OpenAI
 class OpenAI_Parser():
     def __init__(self):
         self.client = OpenAI(api_key = CL.get("openai", "api_key"))
-        self.assistant = self.client.beta.assistants.retrieve(CL.get("openai", "assistant_id"))
+        if not CL.get("openai", "assistant_id") is None: # if assistant_id is set, use it
+            self.assistant = self.client.beta.assistants.retrieve(CL.get("openai", "assistant_id"))
+        else:
+            with open("./src/openai_tools.json", "r") as f:
+                tools = json.load(f)
+            self.assistant = self.client.beta.assistants.create(
+                name="SageCord",
+                instructions = "You are an assistant helping people with their problems. Your response text will be parsed as Markdown format, so please ensure your text output is in standard Markdown.",
+                model = CL.get("openai", "model"),
+                tools = tools
+            )
         # check if thread_mapping exists, else create it
         if os.path.exists("./thread_mapping.json"):
             with open("./thread_mapping.json", "r") as f:
@@ -14,7 +24,10 @@ class OpenAI_Parser():
         else:
             self.thread_mapping_table = {}
 
-    
+    def new_thread(self, context_id):
+        context_id = str(context_id) # convert to string
+        self._clean_expired_threads(context_id)
+
     def get_response(self, content, attachments, context_id):
         context_id = str(context_id) # convert to string
         uploaded_files = self._parse_attachments(attachments)
@@ -22,7 +35,9 @@ class OpenAI_Parser():
         self._set_thread_files(context_id, uploaded_files)
         user_message = self._create_message(content, uploaded_files, thread) # create message containing user's message
         run = self._run_thread(thread) # run thread
-        run_result = self._wait_for_run_finish(run)
+        thread_killed = self._wait_for_run_finish(run, context_id)
+        if thread_killed:
+            return [[{"type": "thread_killed"}]]
         new_messages = self._get_new_messages(thread)
         return new_messages
     
@@ -94,9 +109,6 @@ class OpenAI_Parser():
                     "annotations": annotations
                 })
             elif content_type == "image_file":
-                print(89768757657687)
-                print(content)
-                print(self.client.files.retrieve(content.image_file.file_id))
                 file_name = self.client.files.retrieve(content.image_file.file_id).filename
                 # download file, TODO: Add try except
                 file_api_response = self.client.files.with_raw_response.retrieve_content(content.image_file.file_id)
@@ -112,9 +124,8 @@ class OpenAI_Parser():
                 raise NotImplementedError
         return parsed_contents
 
-
-
-    def _wait_for_run_finish(self, run):
+    # return True if thread is killed, else return False
+    def _wait_for_run_finish(self, run, context_id):
         for _ in range(60 * 2): # 60 seconds
             run = self.client.beta.threads.runs.retrieve( # update run
                 thread_id = run.thread_id,
@@ -131,10 +142,31 @@ class OpenAI_Parser():
                 return
             elif run.status == "cancelled":
                 return
-            elif run.status == "requires_action":
-                return
+            elif run.status == "requires_action": # function calling triggered
+                thread_killed = self._process_function_calling(run, context_id)
+                if thread_killed:
+                    return True # thread is killed
             time.sleep(0.5)
-        return # TODO: timeout
+        return False # TODO: timeout
+    
+    def _process_function_calling(self, run, context_id):
+        thread_id = run.thread_id
+        run_id = run.id
+        # check which function is called
+        tool_outputs = []
+        for func_call in run.required_action.submit_tool_outputs.tool_calls:
+            if func_call.function.name == "new_thread": # new thread
+                self.new_thread(context_id)
+                return True # thread is killed
+        run = self.client.beta.threads.runs.submit_tool_outputs(
+            thread_id = thread_id,
+            run_id = run_id,
+            tool_outputs = tool_outputs
+        )
+        return False # thread is not killed
+
+        
+
 
     def _run_thread(self, thread):
         run = self.client.beta.threads.runs.create(
@@ -170,6 +202,8 @@ class OpenAI_Parser():
             return thread
         
     def _clean_expired_threads(self, context_id):
+        if context_id not in self.thread_mapping_table:
+            return
         for file_id in self.thread_mapping_table[context_id]["file_ids"]:
             self.client.files.delete(file_id)
         self.client.beta.threads.delete(self.thread_mapping_table[context_id]["thread_id"])
